Refactoring: thread safety support + BigDecimal values + Edge case unit tests

+++ src/main/java/com/banking/interfaces/BankAccount.java
@@ -1,9 +1,11 @@
 package com.banking.interfaces;
 
+import java.math.BigDecimal;
+
 public interface BankAccount {
-    int getBalance();
+    BigDecimal getBalance();
 
-    void deposit(int amount);
+    void deposit(BigDecimal amount);
 
-    void withdraw(int amount);
+    void withdraw(BigDecimal amount);
 }
